+++ packages/lumx-react/src/components/slideshow/SlideshowControls.tsx
@@ -3,7 +3,7 @@ import React, { forwardRef, RefObject, useCallback, useMemo } from 'react';
 import classNames from 'classnames';
 import range from 'lodash/range';
 
-import { mdiChevronLeft, mdiChevronRight } from '@lumx/icons';
+import { mdiChevronLeft, mdiChevronRight, mdiPlayCircleOutline, mdiPauseCircleOutline } from '@lumx/icons';
 import { Emphasis, IconButton, IconButtonProps, Theme } from '@lumx/react';
 import { Comp, GenericProps, getRootClassName, handleBasicClasses } from '@lumx/react/utils';
 import { WINDOW } from '@lumx/react/constants';
@@ -38,6 +38,13 @@ export interface SlideshowControlsProps extends GenericProps {
     onPaginationClick?(index: number): void;
     /** On previous button click callback. */
     onPreviousClick?(loopback?: boolean): void;
+    /** whether the slideshow is currently playing */
+    isAutoPlaying?: boolean;
+    /** function to be executed in order to retrieve the label for the pagination item */
+    paginationItemLabel?: (index: number) => string;
+    /** Props to pass to the lay button (minus those already set by the SlideshowControls props). */
+    playButtonProps?: Pick<IconButtonProps, 'label'> &
+        Omit<IconButtonProps, 'label' | 'onClick' | 'icon' | 'emphasis' | 'color'>;
 }
 
 /**
@@ -77,6 +84,9 @@ const InternalSlideshowControls: Comp<SlideshowControlsProps, HTMLDivElement> = 
         previousButtonProps,
         slidesCount,
         theme,
+        isAutoPlaying = false,
+        playButtonProps,
+        paginationItemLabel,
         ...forwardedProps
     } = props;
 
@@ -130,7 +140,6 @@ const InternalSlideshowControls: Comp<SlideshowControlsProps, HTMLDivElement> = 
                                 const isActive = activeIndex === index;
                                 const isOutRange = index < visibleRange.min || index > visibleRange.max;
                                 return (
-                                    // eslint-disable-next-line jsx-a11y/control-has-associated-label
                                     <button
                                         className={classNames(
                                             handleBasicClasses({
@@ -143,13 +152,36 @@ const InternalSlideshowControls: Comp<SlideshowControlsProps, HTMLDivElement> = 
                                         key={index}
                                         type="button"
                                         onClick={() => onPaginationClick?.(index)}
+                                        {...{
+                                            'aria-label': paginationItemLabel
+                                                ? paginationItemLabel(index)
+                                                : `${index + 1} / ${slidesCount}`,
+                                        }}
                                     />
                                 );
                             }),
-                        [slidesCount, visibleRange.min, visibleRange.max, activeIndex, onPaginationClick],
+                        [
+                            slidesCount,
+                            visibleRange.min,
+                            visibleRange.max,
+                            activeIndex,
+                            paginationItemLabel,
+                            onPaginationClick,
+                        ],
                     )}
                 </div>
             </div>
+
+            {playButtonProps ? (
+                <IconButton
+                    {...playButtonProps}
+                    icon={isAutoPlaying ? mdiPauseCircleOutline : mdiPlayCircleOutline}
+                    className={`${CLASSNAME}__play`}
+                    color={theme === Theme.dark ? 'light' : 'dark'}
+                    emphasis={Emphasis.low}
+                />
+            ) : null}
+
             <IconButton
                 {...nextButtonProps}
                 icon={mdiChevronRight}