+++ packages/lumx-react/src/components/slideshow/index.ts
@@ -1,3 +1,4 @@
 export * from './Slideshow';
 export * from './SlideshowItem';
 export * from './SlideshowControls';
+export * from './Slides';